requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
lxml==4.9.1
multitasking==0.0.11
mypy-extensions==0.4.3
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pathspec==0.10.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
yfinance==0.1.74
